move subscription repositories into a repo folder
