+++ PubSub.Application/Services/SubscriptionService.cs
@@ -3,6 +3,7 @@ using PubSub.Application.Models;
 using PubSub.Core.Abstractions.ErrorTypes;
 using PubSub.Core.Models;
 using PubSub.DataAccess;
+using PubSub.DataAccess.Repositories;
 
 namespace PubSub.Application.Services;
