+++ PubSub.Test/Unit/Application/SubscriptionServiceTest.cs
@@ -5,6 +5,7 @@ using PubSub.Application.Services;
 using PubSub.Core.Abstractions.ErrorTypes;
 using PubSub.Core.Models;
 using PubSub.DataAccess;
+using PubSub.DataAccess.Repositories;
 using PubSub.Test.Base.Tests.Toolkit.Patterns;
 
 namespace PubSub.Test.Unit.Application
